a module for zypp::Zypp
so that its contents are private
